python/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.1.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
bitarray==3.0.0
blinker==1.9.0
Brotli==1.0.9
cached-property==2.0.1
capstone==6.0.0a2
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.7
cloud-init==24.4
colorama==0.4.4
colored-traceback==0.4.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cors==1.0.1
cryptography==3.4.8
cycler==0.11.0
cytoolz==1.0.1
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.0
eth-keys==0.6.1
eth-rlp==1.0.1
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
filelock==3.17.0
Flask==3.0.2
Flask-Cors==5.0.0
fonttools==4.29.1
fpdf==1.7.2
frozenlist==1.5.0
fs==2.4.12
future==1.0.0
gast==0.5.2
gevent==24.11.1
greenlet==3.1.1
hexbytes==0.3.1
hexdump==3.3
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
intervaltree==3.1.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lru-dict==1.3.0
lxml==4.8.0
lz4==3.1.3+dfsg
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.5.1
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.1.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
parsimonious==0.10.0
patsy==1.0.1
pbkdf2==1.3
Pillow==9.0.1
plumbum==1.9.0
ply==3.11
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.0
pwntools==4.13.1
py-ecc==7.0.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pycurl==7.44.1
pyelftools==0.31
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pythran==0.10.0
pytz==2022.1
pyunormalize==16.0.0
PyYAML==5.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
rlp==4.0.1
ROPGadget==7.5
rpds-py==0.22.3
rpyc==6.0.1
scipy==1.8.0
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.1
statsmodels==0.14.4
sympy==1.9
systemd-python==234
tink==1.10.0
tldextract==5.1.3
toolz==1.0.0
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unicorn==2.1.1
unix-ar==0.2.1
urllib3==1.26.5
wadllib==1.3.6
web3==6.11.1
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
zipp==1.0.0
zope.event==5.0
zope.interface==5.4.0
zstandard==0.23.0
